### 2024-05-17 #6761 @editorialbot
[PRE REVIEW]: FastVPINNs: An efficient tensor-based Python library for solving partial differential equations using hp-Variational Physics Informed Neural Networks
**Submitting author:** @thivinanandh (<a href="http://orcid.org/0000-0003-4969-3242">Thivin Anandh</a>)
**Repository:** https://github.com/cmgcds/fastvpinns
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.1
