### 2024-05-17 #6761 @danielskatz
👋 @thivinanandh - Thanks for your submission. I'll now look for an editor.

While I do this, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
