+++ joss/paper.bib
@@ -5,7 +5,8 @@
   volume={378},
   pages={686--707},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.jcp.2018.10.045}
 }
 
 @misc{hp_vpinns_github,
@@ -24,7 +25,8 @@
       year={2024},
       eprint={2404.12063},
       archivePrefix={arXiv},
-      primaryClass={cs.LG}
+      primaryClass={cs.LG},
+      doi={https://doi.org/10.48550/arXiv.2404.12063}
 }
 
 @article{kharazmi2021hp,
@@ -34,13 +36,15 @@
   volume={374},
   pages={113547},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.cma.2020.113547}
 }
 @article{kharazmi2019variational,
   title={{Variational physics-informed neural networks for solving partial differential equations}},
   author={Kharazmi, Ehsan and Zhang, Zhongqiang and Karniadakis, George Em},
   journal={arXiv preprint arXiv:1912.00873},
-  year={2019}
+  year={2019},
+  doi={https://doi.org/10.48550/arXiv.1912.00873}
 }
 
 @article{tensorflow2015-whitepaper,
@@ -89,11 +93,13 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
+  doi={https://doi.org/10.48550/arXiv.1603.04467}
 }
 
 @book{ganesan2017finite,
   title={Finite elements: Theory and algorithms},
   author={Ganesan, Sashikumaar and Tobiska, Lutz},
   year={2017},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={https://doi.org/10.1017/9781108235013}
 }
