+++ joss/paper.bib
@@ -6,7 +6,7 @@
   pages={686--707},
   year={2019},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.jcp.2018.10.045}
+  doi = {10.1016/j.jcp.2018.10.045}
 }
 
 @misc{hp_vpinns_github,
@@ -26,7 +26,7 @@
       eprint={2404.12063},
       archivePrefix={arXiv},
       primaryClass={cs.LG},
-      doi={https://doi.org/10.48550/arXiv.2404.12063}
+      doi={10.48550/arXiv.2404.12063}
 }
 
 @article{kharazmi2021hp,
@@ -37,14 +37,14 @@
   pages={113547},
   year={2021},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.cma.2020.113547}
+  doi={10.1016/j.cma.2020.113547}
 }
 @article{kharazmi2019variational,
   title={{Variational physics-informed neural networks for solving partial differential equations}},
   author={Kharazmi, Ehsan and Zhang, Zhongqiang and Karniadakis, George Em},
   journal={arXiv preprint arXiv:1912.00873},
   year={2019},
-  doi={https://doi.org/10.48550/arXiv.1912.00873}
+  doi={10.48550/arXiv.1912.00873}
 }
 
 @article{tensorflow2015-whitepaper,
@@ -93,7 +93,7 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
-  doi={https://doi.org/10.48550/arXiv.1603.04467}
+  doi={10.48550/arXiv.1603.04467}
 }
 
 @book{ganesan2017finite,
@@ -101,5 +101,5 @@ author={
   author={Ganesan, Sashikumaar and Tobiska, Lutz},
   year={2017},
   publisher={Cambridge University Press},
-  doi={https://doi.org/10.1017/9781108235013}
+  doi={10.1017/9781108235013}
 }
