### 2024-05-17 #6761 @thivinanandh
@danielskatz , Sorry for the delay in response. I have fixed the DOI errors by adding DOI references to the bib file. The one error comes from the bib item which refers  to the hp-VPINNs github repository, which does not have any DOI.
