### 2024-05-19 #6764 @editorialbot
[REVIEW]: FastVPINNs: An efficient tensor-based Python library for solving partial differential equations using hp-Variational Physics Informed Neural Networks
**Submitting author:** @thivinanandh (<a href="http://orcid.org/0000-0003-4969-3242">Thivin Anandh</a>)
**Repository:** https://github.com/cmgcds/fastvpinns
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.0.2
