+++ README.md
@@ -8,7 +8,7 @@
 [![MIT License](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 ![Python Versions](https://img.shields.io/badge/python-3.8%20|%203.9%20|%203.10%20|%203.11-blue)
-
+[![status](https://joss.theoj.org/papers/173453d08813f66a718f583502351747/status.svg)](https://joss.theoj.org/papers/173453d08813f66a718f583502351747)
 
 
 <br />
