### 2024-05-23 d62de2c
Minor corrections for the README file.
