### 2024-05-23 df3853d
Minor corrections to README file.
