+++ fastvpinns/utils/compute_utils.py
@@ -1,14 +1,14 @@
 """
-filename: compute_utils.py
-description: This file contains the utility functions for 
-             computing the errors between the exact and 
-             predicted solutions
-author: Thivin Anandh D
-date: 02/11/2023
-changelog: 02/11/2023 - file created
-           02/11/2023 - added functions to compute L1, L2, L_inf errors
+This file `compute_utils.py` contains the utility functions for 
+computing the errors between the exact and predicted solutions.
 
-known_issues: None
+Author: Thivin Anandh D
+
+Date: 02/Nov/2023
+
+Changelog: 02/Nov/2023 - file created; and added functions to compute L1, L2, L_inf errors
+
+Known issues: None
 """
 
 # Importing the required libraries
@@ -16,19 +16,23 @@ import numpy as np
 
 
 def compute_l2_error(u_exact, u_approx):
-    """This function will compute the L2 error between the exact solution and the approximate solution.
-    The L2 error is defined as:
-
-    ..math::
-        \\sqrt{\\frac{1}{N} \\sum_{i=1}^{N} (u_{exact} - u_{approx})^2}
+    """
+    This function will compute the L2 error between the exact solution and the approximate solution.
 
     :param u_exact: numpy array containing the exact solution
     :type u_exact: numpy.ndarray
     :param u_approx: numpy array containing the approximate solution
     :type u_approx: numpy.ndarray
+    
     :return: L2 error between the exact and approximate solutions
     :rtype: float
     """
+    
+    #The L2 error is defined as:
+
+    #   ..math::
+    #        \\sqrt{\\frac{1}{N} \\sum_{i=1}^{N} (u_{exact} - u_{approx})^2}
+
     # Flatten the arrays
     u_exact = u_exact.flatten()
     u_approx = u_approx.flatten()
@@ -39,17 +43,23 @@ def compute_l2_error(u_exact, u_approx):
 
 
 def compute_l1_error(u_exact, u_approx):
-    """This function will compute the L1 error between the exact solution and the approximate solution.
-    The L1 error is defined as:
-    ..math::
-        \\frac{1}{N} \\sum_{i=1}^{N} |u_{exact} - u_{approx}|
+    """
+    This function will compute the L1 error between the exact solution and the approximate solution.  
+    
     :param u_exact: numpy array containing the exact solution
     :type u_exact: numpy.ndarray
     :param u_approx: numpy array containing the approximate solution
     :type u_approx: numpy.ndarray
+    
     :return: L1 error between the exact and approximate solutions
     :rtype: float
     """
+
+    #The L1 error is defined as:
+
+    #    ..math::
+    #        \\frac{1}{N} \\sum_{i=1}^{N} |u_{exact} - u_{approx}|
+
     # Flatten the arrays
     u_exact = u_exact.flatten()
     u_approx = u_approx.flatten()
@@ -59,17 +69,23 @@ def compute_l1_error(u_exact, u_approx):
 
 
 def compute_linf_error(u_exact, u_approx):
-    """This function will compute the L_inf error between the exact solution and the approximate solution.
-    The L_inf error is defined as
-        ..math::
-            \\max_{i=1}^{N} |u_{exact} - u_{approx}|
+    """
+    This function will compute the L_inf error between the exact solution and the approximate solution.
+    
     :param u_exact: numpy array containing the exact solution
     :type u_exact: numpy.ndarray
     :param u_approx: numpy array containing the approximate solution
     :type u_approx: numpy.ndarray
+    
     :return: L_inf error between the exact and approximate solutions
     :rtype: float
     """
[…]
