### 2024-07-08 cmgcds/fastvpinns#15 @pescap
Installation and reproducibility
Dear all,
I successfully installed the library from the [DeepXde Docker image](https://hub.docker.com/r/pescapil/deepxde), which includes TensorFlow 2.9.2.

For your information, I had to further install the following libraries:

```
apt-get install libgl1-mesa-glx
apt-get install libglu1
apt-get install libxcursor-dev
apt-get install libxft2
apt-get install libxinerama
```

Is it a known behavior?

It could be beneficial to build a dedicated image for fastvpinns, for reproducibility purposes.
I could help you with this.

Finally, do you recommend any version of TensorFlow?
Best,
