### 2024-07-08 #6764 @pescap
Dear all, thank you for this very interesting and useful library. I have a few recommendations:
- [Documentation](https://cmgcds.github.io/fastvpinns/index.html): Add a more thorough statement of need. It could be a summarized version of the statement of need you propose in the software paper.
- Reproducibility: I created an [issue](https://github.com/cmgcds/fastvpinns/issues/15) sharing my experience installing the library to improve its reproducibility even more.
- Software paper quality of writing: The quality could be improved. I recommend that you recheck the writing and formulas.
