### 2024-07-09 d93f7df
Added Docker file
A Docker image for fastvpinns based on cuda 11.2 , cudnn 8.1 and tensorflow 2.11
Downgraded the tensorflow version to 2.11 to match with the cudnn compatibility
Docker image uploaded to thivinanandh/fastvpinns
