+++ docs/index.rst
@@ -33,15 +33,20 @@ Welcome to FastVPINNs's documentation!
 .. image:: https://img.shields.io/badge/python-3.8%20|%203.9%20|%203.10%20|%203.11-blue
    :alt: Python Versions
 
-A robust tensor-based deep learning framework for solving PDE's using hp-Variational Physics-Informed Neural Networks (hp-VPINNs). The framework supports handling complex geometries and uses tensor-based loss computation to accelerate the training of conventional hp-VPINNs. 
-The framework is based on the work by `FastVPINNs Paper <https://arxiv.org/abs/2404.12063>`_. The framework is written on `Tensorflow 2.0 <https://www.tensorflow.org/>`_ and has support for handling external meshes. 
 
-*Note: This framework is a highly optimised version of the the initial implementation of hp-VPINNs by* `kharazmi <https://github.com/ehsankharazmi/hp-VPINNs>`_. Ref `hp-VPINNs(arXiv) <https://arxiv.org/abs/2003.05385>`_.
 
-Variational Physics-informed neural network
+Statement of Need
+~~~~~~~~~~~~~~~~~
+A robust tensor-based deep learning framework for solving PDEs using hp-Variational Physics-Informed Neural Networks (hp-VPINNs). The framework is written on `Tensorflow 2.0 <https://www.tensorflow.org/>`_ and supports handling external meshes. This framework is a highly optimized version of the initial implementation of hp-VPINNs by `kharazmi <https://github.com/ehsankharazmi/hp-VPINNs>`_. Refer to `hp-VPINNs(arXiv) <https://arxiv.org/abs/2003.05385>`_. The framework uses tensor-based loss computation to accelerate the training of conventional hp-VPINNs by 100X for domains with a large number of elements. Furthermore, the library supports reading external meshes (in .mesh format), which can be used to solve problems in real-world scenarios. The library provides high-level functional APIs to define the test functions and tensor assembly, allowing users to train an hp-VPINNs problem with fewer than 6 API calls.
+
+The framework is based on the work by `FastVPINNs Paper <https://arxiv.org/abs/2404.12063>`_. 
+
+
+
+hp-Variational Physics-informed neural network
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Variational Physics-Informed neural networks are special form of physics informed neural networks, which uses variational form of the loss function to train the NN. A special form of hp-Variational PINNs which uses h- & p- refinement to enhance the ability of the NN to capture higher frequency solutions. 
+hp-Variational Physics-Informed neural networks are special form of physics informed neural networks, which uses variational form of the loss function to train the NN. A special form of hp-Variational PINNs which uses h- & p- refinement to enhance the ability of the NN to capture higher frequency solutions. 
 For more details on the theory and implementation of hp-VPINNs, please refer to the `FastVPINNs Paper <https://arxiv.org/abs/2404.12063>`_ and `hp-VPINNs Paper <https://arxiv.org/abs/2003.05385>`_.
 
 .. include an image here
