### 2024-07-09 2433468
Removed MacOS compatibility for the lower version of the tensorflow
