+++ .github/workflows/compatibility-tests.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, ubuntu-20.04, macos-latest, windows-latest]
+        os: [ubuntu-latest, ubuntu-20.04, windows-latest]
         python-version: ['3.8', '3.9', '3.10', '3.11']
 
     steps:
