### 2024-07-09 2c426a5
Reverted back to tensorflow version 2.13 and added checks for MacOS
