### 2024-07-09 dc7858a
Added exception handling
Added try statements for CPU only executions
