### 2024-07-09 454de5b
Docker changes 1, Added a Sed command to downgrade tensorflow version to 2.11 2, Fixed the tnvcc to nvcc bug
