+++ Dockerfile
@@ -30,6 +30,9 @@ RUN apt-get update && apt-get install -y \
 # git clone the official repository of fastvpinns
 RUN git clone https://github.com/cmgcds/fastvpinns.git
 
+# open the requirements.txt file and replace the "tensorflow>=2.9.1,<=2.13.0" with "tensorflow>=2.9.1,<=2.11.0"
+RUN sed -i 's/tensorflow>=2.9.1,<=2.13.0/tensorflow>=2.9.1,<=2.11.0/g' /fastvpinns/requirements.txt
+
 # used for debugging
 # COPY requirements.txt /fastvpinns/requirements.txt 
 
@@ -49,14 +52,14 @@ RUN echo "export PYTHONPATH=${PYTHONPATH}:/fastvpinns" >> ~/.bashrc
 RUN echo "alias python=python3" >> ~/.bashrc
 
 # temporarily copy the file
-# COPY docker_initialise.py /docker_initialise.py
+COPY docker_initialise.py /fastvpinns/docker_initialise.py
 # Change execution permission
-RUN chmod +x /docker_initialise.py
+RUN chmod +x /fastvpinns/docker_initialise.py
 
 
 # set bash as the default command
 CMD ["/bin/bash"]
 
 
-ENTRYPOINT ["bash", "-c", "python3 /docker_initialise.py && exec /bin/bash"]
+ENTRYPOINT ["bash", "-c", "python3 /fastvpinns/docker_initialise.py && exec /bin/bash"]
 
