### 2024-07-09 cmgcds/fastvpinns#15 @thivinanandh
Hi @pescap ,

Thanks for raising a request for the enhancement.

### Additional requirements :

Yes as you have mentioned `libgl1-mesa-glx` library will be needed as additional installation, which was mentioned in the installation section of the documentation [Link](https://cmgcds.github.io/fastvpinns/_rst/_installation.html). However, in our tests we did not find the need to install other libraries, which were mentioned above. If you could provide us an example where `libgl1-mesa-glx` is not sufficient, then we can go ahead and add the additional libraries to the documentation page.

### Tensorflow Version
Currently the code is tested upto tensorflow version 2.13 (as observed from the requirements.txt file).
. Depending on your OS, cuDNN, and CUDA version, any version of TensorFlow from 2.9 to 2.13 can be used.

### Dedicated Docker Image
As suggested, I have created a new docker image at the url [FastPINNs-Docker Link](https://hub.docker.com/repository/docker/thivinanandh/fastvpinns),You can pull the image using the command:

 `docker pull thivinanandh/fastvpinns:latest`.

The docker image will have the welcome message, which displays the cuda driver, gpu and tensorflow version

![image](https://github.com/cmgcds/fastvpinns/assets/52538701/1cf3492f-7989-4354-a57a-1e3c755b1828)

The Docker image is built on Ubuntu 20.04, with CUDA version 11.2 and cuDNN version 8.1, and it includes TensorFlow version 2.11 (since tensorflow 2.11 is compiled with cuDNN 8.1). The additional documentation on how to use the docker image is available at the [FastVPINNs-Docker Documentation Page](https://cmgcds.github.io/fastvpinns/_rst/_docker.html)

Please let me know if you need any additional information regarding the same
