### 2024-07-11 cmgcds/fastvpinns#15 @pescap
Dear all, great!

When I run the container I get:

````
docker run --gpus all -it --rm -v ~/fastvpinns:/fastvpinns thivinanandh/fastvpinns:latest
Traceback (most recent call last):
  File "/docker_initialise.py", line 85, in <module>
    main()
  File "/docker_initialise.py", line 50, in main
    version = get_version_from_toml()
  File "/docker_initialise.py", line 6, in get_version_from_toml
    with open("pyproject.toml", "r") as file:
FileNotFoundError: [Errno 2] No such file or directory: 'pyproject.toml'
