### 2024-07-15 cmgcds/fastvpinns#15 @pescap
> Hi @pescap ,
>
> Bug: ` [Errno 2] No such file or directory: 'pyproject.toml'`
>
> Root Cause : It seems that in the `Dockerfile` i have used a git clone command to pull the codes from the github repository to get the updated codes and files. However , during my build, docker seems to have used a cached version of the `git clone` command and the recent changes made to the codes were not reflected in the local build.
>
> Fix: The issue is fixed now, by running the build using `--no-cache flag` .
>
> Could you kindly try once again to see if the issue persists or not.

Hi, has the Docker image been updated [here](https://hub.docker.com/r/thivinanandh/fastvpinns)? (with docker push)
