### 2024-07-17 cmgcds/fastvpinns#15 @thivinanandh
@pescap , yes it has been updated with docker push. The Current image on dockerhub is the updated image

Thanks
