### 2024-07-17 #6764 @thivinanandh
Hi @diehlpk ,
Thank you for the gentle reminder.

The Reproducibility issues raised by @pescap are currently being addressed in the [FastVPINNs-Github issues](https://github.com/cmgcds/fastvpinns/issues/15).

- Documentation : A proper statement of need is added to the documentation
- Quality of Writting : @pescap , we are looking into the quality of writing. Could you kindly let us know are there any specific issues that stood out that you want us to address ( in terms of both writing and formulas ). It would be much helpful for us to make these corrections.
