### 2024-07-24 cmgcds/fastvpinns#17 @ziyiyin97
JOSS review
First of all, I sincerely apologize for my late reviews. I encountered some unforeseeable issues recently, which delayed my response to your manuscript. Thank you again for understanding.

Overall, it is an awesome software package dedicated to hp-VPINNs. The statement of need is clear and well-motivated to me. The tutorials are helpful and the API documents are thorough.

The only concerns I have are:
1. #16 I believe the tutorial docs should be fixed this way but correct me if wrong
2. One of the claims in the manuscript is that this package can read external mesh from a `.mesh` file. This is substantiated in the tutorial scripts as they indeed load mesh from a `.mesh` file. However, it is not so clear how a `.mesh` file is generated from GMSH package. I suggest a short tutorial script on how to generate a `.mesh` file from a user-defined mesh, especially one that is slightly more "complicated" or "arbitrary" than uniform mesh or circular mesh. With this tutorial, one could examine the applicability of this package to more realistic problems at hand where the mesh is somehow irregular.
