### 2024-07-24 cmgcds/fastvpinns#17 @divijghose
Hello @ziyiyin97,
Thank you for your review of our work. We have tried to address the concerns you have raised as follows:
1. The minor errors in documentations have been fixed by merging your changes #16  into main.
2. We have added two tutorials on your documentation site that deal with mesh generation using Gmsh. One deals with [mesh generation using external CAD files](https://cmgcds.github.io/fastvpinns/_rst/tutorials/mesh_generation/mesh.html) and the other with [mesh generation from GEO files](https://cmgcds.github.io/fastvpinns/_rst/tutorials/mesh_generation/mesh.html#creating-the-geometry-in-geo-file). With these tutorials, a user should be able to leverage the external mesh capabilities of FastVPINNs.
![image](https://github.com/user-attachments/assets/77d39d4e-0e56-47cf-84c2-4d4442fb26a5)

If you are satisfied with these changes, I can close this issue and you can do the needful for the JOSS review. If not, do get back to us.
