### 2024-07-24 685d16e
Docker Image fix 1, Adhoc measures to copy the recent code to the Docker image, rather than pulling it from the actual git repo 2, Changed documentation to reflect proper mounting of an external folder to an internal output folder rather than the root folder. An issue was raised, when the previous command given in the documentation tried to mount an empty external folder to the internal main folder, which was causing the issues. Now the documentation is updated with the proper command
