### 2024-07-24 cmgcds/fastvpinns#15 @thivinanandh
Hi @pescap ,

I was attending a conference past week thats why i was not able to reply back in time.  Please find below the fix for the issue reported with docker.

**Issue**

File not found while trying to launch the docker terminal

```bash
[Errno 2] No such file or directory: 'pyproject.toml'
```

**Root Cause**

It seems there was no issue with the docker build, rather the issue was with the usage of `-v` flag which we used to mount an external volume. When we use `-v ~/fastvpinns:/fastvpinns`, we were mounting an empty directory from a host named `~/fastvpinns` to the actual directory  within the docker container `/fastvpinns`, which contains the code base. So when this command is executed, the empty directory overwrites the core directory `/fastvpinns` within our docker environment causing the docker build to look for a file, which has been deleted now due to the mounting.

**Fix**
The fix is to mount an external directory from host into a new directory such as `/fastvpinns/output` within the docker environment, so that it does not overwrite the existing codes. Updated command for this is as follows,

```bash
docker run --gpus all -it --rm -v ~/fastvpinns_docker_output:/fastvpinns/output thivinanandh/fastvpinns:latest
```

**Additional Update**

The explanation for the same is now added to the docker documentation page of FastVPINNs and the running instructions are updated accordingly. [Docker Documentation - FastVPINNs](https://cmgcds.github.io/fastvpinns/_rst/_docker.html)
