### 2024-07-24 #6764 @thivinanandh
@ziyiyin97
We have addressed the issues on [cmgcds/fastvpinns#17](https://github.com/cmgcds/fastvpinns/issues/17).
-  We had approved the PR and merged the changes to the documentation as proposed.
- We have also added a separate documentation on how to generate external meshes which are compatible with FastVPINNs at [Mesh Generation - FastVPINNs Documentation](https://cmgcds.github.io/fastvpinns/_rst/tutorials/mesh_generation/mesh.html)

Please let us know if you need any additional update from our side.
