### 2024-07-25 6e5195f
Modifications to Joss Paper
