+++ joss/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Introduction
 
-Partial differential equations (PDEs) are essential in modeling physical phenomena such as heat transfer, electromagnetics and fluid dynamics. The current progress in the field of scientific machine learning (SciML) has resulted in incorporating deep learning and data-driven methods to solve PDEs. Physics informed neural networks introduced by [@raissi2019physics] works by incorporating the governing equations of the physical systems into the Neural Network's loss function. Let us consider a two-dimensional Poisson equation as an example which reads as follows:
+Partial differential equations (PDEs) are essential in modeling physical phenomena such as heat transfer, electromagnetics and fluid dynamics. The current progress in the field of scientific machine learning (SciML) has resulted in incorporating deep learning and data-driven methods to solve PDEs. Let us consider a two-dimensional Poisson equation as an example, defined on the domain $\Omega$, which reads as follows:
 
 \begin{align}
     - \nabla^2 u(x) &= f(x), \quad \Omega \in \mathbb{R}^2,  \\
@@ -37,23 +37,41 @@ Partial differential equations (PDEs) are essential in modeling physical phenome
 
 Here, $x \in \Omega$ is the spatial co-ordinate, $u(x)$ is the solution of the PDE, $f(x)$ is a known source term and $g(x)$ is the value of the solution on the domain boundary, $\partial \Omega$. 
 
+A neural network is a parametric function of $x$, denoted as $u_{\text{NN}}(x; W, b)$. In this context, $W$ and $b$ represent the weights and biases of the network. When the neural network consists of $h$ hidden layers, with the $i^{\text{th}}$ layer containing $n_i$ neurons, the mathematical representation of the function takes the following shape:
+\begin{equation*}
+   u_{\text{NN}}(x; W, b) = l \circ \mathrm{T}^{(h)} \circ \mathrm{T}^{(h-1)} \hdots \circ \mathrm{T}^1(x)
+    \label{eq:NN}
+\end{equation*}
+
+Here, $l : \mathbb{R}^{n_h} \rightarrow \mathbb{R}$ is a linear mapping in the output layer and $\mathrm{T}^{(i)}(\cdot) = \sigma(W^{(i)} \times \cdot + b^{(i)})$ is a non-linear mapping in the $i^{th}$ layer $( i= 1, 2, \cdots , h)$, with the non-linear activation function $\sigma$ and the weights $W^{(i)}$ and biases $b^{(i)}$ of the respective layers.
+
 Physics-informed neural networks (PINNs), introduced by [@raissi2019physics], work by incorporating the governing equations of the physical systems and the boundary conditions as physics-based constrains to train the neural networks. The core idea of PINNs lies in the ability of obtaining the gradients of the solutions with respect to the input using the automatic differention routines available within deep learning frameworks such as tensorflow [@tensorflow2015-whitepaper]. The loss function for the PINNs can be written as follows:
 
 \begin{align*}
     \begin{split}
-        L_p(W,b) &= \frac{1}{N_T}\sum_{t=1}^{N_T}\left|(-\Delta u_{\text{NN}}(x_t;W,b) - f(x_t))\right|^2,\\
+        L_p(W,b) &= \frac{1}{N_T}\sum_{t=1}^{N_T}\left|(-\nabla^2 u_{\text{NN}}(x_t;W,b) - f(x_t))\right|^2,\\
         L_b(W,b) &= \frac{1}{N_D}\sum_{d=1}^{N_{D}}\left|u_{\text{NN}}(x_d; W, b) - g(x_d)\right|^2,\\
         L_{\text{PINN}}(W,b) &= L_p + \tau L_b.
     \end{split}
 \end{align*}
 
[…]
