### 2024-07-25 #6764 @thivinanandh
> Dear all, thank you for this very interesting and useful library. I have a few recommendations:
>
> * [Documentation](https://cmgcds.github.io/fastvpinns/index.html): Add a more thorough statement of need. It could be a summarized version of the statement of need you propose in the software paper.
> * Reproducibility: I created an [issue](https://github.com/cmgcds/fastvpinns/issues/15) sharing my experience installing the library to improve its reproducibility even more.
> * Software paper quality of writing: The quality could be improved. I recommend that you recheck the writing and formulas.

@pescap , we have addressed your issues mentioned, please find below the overview

- **Documentation** : A proper statement of need is added to the documentation page as suggested
- **Reproducibility** : We have addressed the issues with the docker image and have updated our comments in the [issue link](https://github.com/cmgcds/fastvpinns/issues/15)
- **Software paper quality of writing** : We have rewritten the contents of the paper and corrected some of the Mathematical expressions.

Kindly update your comments on the issue page, so that we can close the issue

Please let us know if you need any additional update from our side
