### 2024-07-25 cmgcds/fastvpinns#17 @ziyiyin97
I think these changes are great!
