### 2024-07-26 #6764 @thivinanandh
We would like to thank @pescap and @ziyiyin97 for taking time of their schedule and providing their valuable feedbacks and corrections to the project. Both the docker update and the mesh generation documentation seemed to be a valuable addition to the project.

@diehlpk , could you kindly guide us on what needs to be done next from our side, given that both the reviewers have completed their checklists ?
