### 2024-07-26 85b5b32
Joss paper Merge (#19)
Squashing and merging all the changes from the joss paper branch into a single commit on main branch

* Joss paper

Added  for the paper
Added  for the references
Added  in the github workflowa

* Joss CI bug Fix

FIxed the path on the joss-paper.yml

* Removed Older Joss Paper folder

* Added name to Joss PDF CI

* Fixed broken citations

* Removed the compatibility test for push to other branches

* Changes to Joss Paper

Changed Module name from FE2D to FE

* Changed PNG Errors

* Paper corrections

* Paper Correction

* Paper corrections - 3

* Paper corrections - 4

[…]
