+++ joss/paper.md
@@ -1,0 +1,189 @@
+---
+title: 'FastVPINNs: An efficient tensor-based Python library for solving partial differential equations using hp-Variational Physics Informed Neural Networks'
+tags:
+  - python
+  - physics-informed neural networks
+  - scientific machine learning
+  - partial differential equations
+  - hp-variational physics informed neural networks
+  - tensorflow
+authors:
+  - name: Thivin Anandh
+    orcid: 0000-0003-4969-3242
+    affiliation: "1"
+    corresponding: true # corresponding author
+  - name: Divij Ghose
+    orcid: 0009-0005-6295-543X
+    affiliation: "1"
+  - name: Sashikumaar Ganesan
+    orcid: 0000-0003-1858-3972
+    affiliation: "1"
+    corresponding: true # corresponding author
+affiliations:
+ - name: Department of Computational and Data Sciences, Indian Institute of Science, Bangalore, India
+   index: 1
+date: 17 May 2023
+bibliography: paper.bib
+---
+
+# Introduction
+
+Partial differential equations (PDEs) are essential in modeling physical phenomena such as heat transfer, electromagnetics and fluid dynamics. The current progress in the field of scientific machine learning (SciML) has resulted in incorporating deep learning and data-driven methods to solve PDEs. Let us consider a two-dimensional Poisson equation as an example, defined on the domain $\Omega$, which reads as follows:
+
+\begin{align}
+    - \nabla^2 u(x) &= f(x), \quad \Omega \in \mathbb{R}^2,  \\
+    u(x) &= g(x), \quad x \in \partial \Omega.
+\end{align}
+
+Here, $x \in \Omega$ is the spatial co-ordinate, $u(x)$ is the solution of the PDE, $f(x)$ is a known source term and $g(x)$ is the value of the solution on the domain boundary, $\partial \Omega$. 
+
+A neural network is a parametric function of $x$, denoted as $u_{\text{NN}}(x; W, b)$. In this context, $W$ and $b$ represent the weights and biases of the network. When the neural network consists of $h$ hidden layers, with the $i^{\text{th}}$ layer containing $n_i$ neurons, the mathematical representation of the function takes the following shape:
+\begin{equation*}
+   u_{\text{NN}}(x; W, b) = l \circ \mathrm{T}^{(h)} \circ \mathrm{T}^{(h-1)} \hdots \circ \mathrm{T}^1(x)
+    \label{eq:NN}
+\end{equation*}
+
+Here, $l : \mathbb{R}^{n_h} \rightarrow \mathbb{R}$ is a linear mapping in the output layer and $\mathrm{T}^{(i)}(\cdot) = \sigma(W^{(i)} \times \cdot + b^{(i)})$ is a non-linear mapping in the $i^{th}$ layer $( i= 1, 2, \cdots , h)$, with the non-linear activation function $\sigma$ and the weights $W^{(i)}$ and biases $b^{(i)}$ of the respective layers.
+
+Physics-informed neural networks (PINNs), introduced by [@raissi2019physics], work by incorporating the governing equations of the physical systems and the boundary conditions as physics-based constrains to train the neural networks. The core idea of PINNs lies in the ability of obtaining the gradients of the solutions with respect to the input using the automatic differention routines available within deep learning frameworks such as tensorflow [@tensorflow2015-whitepaper]. The loss function for the PINNs can be written as follows:
+
+\begin{align*}
+    \begin{split}
[…]
