+++ joss/paper.bib
@@ -1,0 +1,105 @@
+@article{raissi2019physics,
+  title={Physics-informed neural networks: A deep learning framework for solving forward and inverse problems involving nonlinear partial differential equations},
+  author={Raissi, Maziar and Perdikaris, Paris and Karniadakis, George E},
+  journal={Journal of Computational physics},
+  volume={378},
+  pages={686--707},
+  year={2019},
+  publisher={Elsevier},
+  doi = {10.1016/j.jcp.2018.10.045}
+}
+
+@misc{hp_vpinns_github,
+  author = {Kharazmi, Ehsan},
+  title = {{hp-VPINNs}: High-Performance Variational Physics-Informed Neural Networks},
+  year = {2023},
+  url = {https://github.com/ehsankharazmi/hp-VPINNs},
+  note = {Accessed: 2023-12-01}
+}
+
+@misc{anandh2024fastvpinns,
+      title={FastVPINNs: Tensor-Driven Acceleration
+             of VPINNs for Complex Geometries}, 
+      author={Anandh, Thivin and Ghose, Divij and Jain, Himanshu
+               and Ganesan, Sashikumaar},
+      year={2024},
+      eprint={2404.12063},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      doi={10.48550/arXiv.2404.12063}
+}
+
+@article{kharazmi2021hp,
+  title={{hp-VPINNs: Variational physics-informed neural networks with domain decomposition}},
+  author={Kharazmi, Ehsan and Zhang, Zhongqiang and Karniadakis, George Em},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={374},
+  pages={113547},
+  year={2021},
+  publisher={Elsevier},
+  doi={10.1016/j.cma.2020.113547}
+}
+@article{kharazmi2019variational,
+  title={{Variational physics-informed neural networks for solving partial differential equations}},
+  author={Kharazmi, Ehsan and Zhang, Zhongqiang and Karniadakis, George Em},
+  journal={arXiv preprint arXiv:1912.00873},
+  year={2019},
+  doi={10.48550/arXiv.1912.00873}
+}
+
+@article{tensorflow2015-whitepaper,
+title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+url={https://www.tensorflow.org/},
+note={Software available from tensorflow.org},
+author={
+    Mart\'{i}n~Abadi and
+    Ashish~Agarwal and
+    Paul~Barham and
+    Eugene~Brevdo and
+    Zhifeng~Chen and
+    Craig~Citro and
+    Greg~S.~Corrado and
+    Andy~Davis and
+    Jeffrey~Dean and
+    Matthieu~Devin and
+    Sanjay~Ghemawat and
+    Ian~Goodfellow and
+    Andrew~Harp and
+    Geoffrey~Irving and
+    Michael~Isard and
+    Yangqing Jia and
+    Rafal~Jozefowicz and
+    Lukasz~Kaiser and
+    Manjunath~Kudlur and
+    Josh~Levenberg and
+    Dandelion~Man\'{e} and
+    Rajat~Monga and
+    Sherry~Moore and
+    Derek~Murray and
+    Chris~Olah and
+    Mike~Schuster and
+    Jonathon~Shlens and
+    Benoit~Steiner and
+    Ilya~Sutskever and
+    Kunal~Talwar and
+    Paul~Tucker and
+    Vincent~Vanhoucke and
+    Vijay~Vasudevan and
+    Fernanda~Vi\'{e}gas and
+    Oriol~Vinyals and
+    Pete~Warden and
[…]
