+++ .github/workflows/compatibility-tests.yml
@@ -2,7 +2,7 @@ name: Compatibility tests
 
 on:
   push:
-    branches: [ '*' ]
+    branches: [ 'main' ]
   pull_request:
     branches: [ main ]
   workflow_dispatch:
