### 2024-07-26 #6764 @thivinanandh
@diehlpk , Thanks for the quick update.!

Just a small update from our side, since the review process was completed from the reviewers, we merged the code from "joss_paper" branch (which was set as the official branch for this review) to "main" branch of the repository.
