### 2024-07-28 #6764 @diehlpk
@thivinanandh

- [ ] please add a space in line 60 between the reference and the word. Please be consistent and always use a space.
- [ ] I would prefer to remove the example code and link to the example's source file. So the reader gets to the updated source code if the API changes in the future.
