### 2024-07-29 6deae68
Joss Paper Corrections - 1
- Added the minimal API documentation to the main documentation page
- Made corrections as per the [review](https://github.com/openjournals/joss-reviews/issues/6764)
