### 2024-07-29 #6764 @thivinanandh
> @thivinanandh

>

>    - [x]  please add a space in line 60 between the reference and the word. Please be consistent and always use a space.

>

>    - [x]  I would prefer to remove the example code and link to the example's source file. So the reader gets to the updated source code if the API changes in the future.

@diehlpk.

we have added the space between the reference and the word at all 9 instances within the paper

We removed the code snippet and added a link to the documentation page, where the code will be updated every time a new change involving the API's are pushed.

Please let us know if any further changes are required.
