### 2024-07-29 #6764 @diehlpk
> """
>
> > @thivinanandh I am missing some related work statements in the paper. Are there other projects implementing similar ML? If so, how do they compare to your code?
>
> Dear @diehlpk,
>
> The related work to this paper, which is the first implementation of hp-VPINNs, is referenced both in the Introduction section and the Statement of Need section of the paper. Our project is also a framework to solve hp-VPINNs; however, our framework addresses the two main limitations that exist with the current implementation:
>
>     * The increase in training time with an increase in the number of elements within the domain
>
>     * The inability of the framework to handle complex meshes with skewed quadrilateral elements
>
>
> Our Framework addresses both these limitations and in addition also provides an exhaustive documentation on how to solve PDEs using our framework and provides abstracted APIs which can train a NN to solve a PDE using hp-VPINNs by using fewer API calls.
>
> **Performance Comparison** Due to our tensor-based approach to compute the variational loss, our framework is close to 100x faster in terms of training time when compared to the existing implementation of hp-VPINNs (as mentioned in the statement of need). The validation for these claims can be found in our [paper](https://arxiv.org/abs/2404.12063) (which is also referenced in the statement of need).
>
> Apart from the initial implementation of hp-VPINNs (cited in the paper), to the authors' knowledge, we could not find any other similar implementation available for solving hp-VPINNs. """

I think it would be good to emphasize this more in the paper that this is the first implementation of this method.
