+++ joss/paper.md
@@ -39,7 +39,7 @@ Here, $x \in \Omega$ is the spatial co-ordinate, $u(x)$ is the solution of the P
 
 A neural network is a parametric function of $x$, denoted as $u_{\text{NN}}(x; W, b)$. In this context, $W$ and $b$ represent the weights and biases of the network. When the neural network consists of $h$ hidden layers, with the $i^{\text{th}}$ layer containing $n_i$ neurons, the mathematical representation of the function takes the following shape:
 \begin{equation*}
-   u_{\text{NN}}(x; W, b) = l \circ \mathrm{T}^{(h)} \circ \mathrm{T}^{(h-1)} \hdots \circ \mathrm{T}^1(x)
+   u_{\text{NN}}(x; W, b) = l \circ \mathrm{T}^{(h)} \circ \mathrm{T}^{(h-1)} \hdots \circ \mathrm{T}^1(x).
     \label{eq:NN}
 \end{equation*}
 
@@ -49,8 +49,8 @@ Physics-informed neural networks (PINNs), introduced by [@raissi2019physics], wo
 
 \begin{align*}
     \begin{split}
-        L_p(W,b) &= \frac{1}{N_T}\sum_{t=1}^{N_T}\left|(-\nabla^2 u_{\text{NN}}(x_t;W,b) - f(x_t))\right|^2,\\
-        L_b(W,b) &= \frac{1}{N_D}\sum_{d=1}^{N_{D}}\left|u_{\text{NN}}(x_d; W, b) - g(x_d)\right|^2,\\
+        L_p(W,b) &= \frac{1}{N_T}\sum_{t=1}^{N_T}\left((-\nabla^2 u_{\text{NN}}(x_t;W,b) - f(x_t))\right)^2,\\
+        L_b(W,b) &= \frac{1}{N_D}\sum_{d=1}^{N_{D}}\left(u_{\text{NN}}(x_d; W, b) - g(x_d)\right)^2,\\
         L_{\text{PINN}}(W,b) &= L_p + \tau L_b.
     \end{split}
 \end{align*}
@@ -68,23 +68,24 @@ Variational Physics informed neural networks (VPINNs) [@kharazmi2019variational]
 \end{equation*}
 
 
-The loss function of hp-VPINNs with $\texttt{N\_{elem}}$ elements in the domain can be written as follows.
+The loss function of hp-VPINNs with $\texttt{N\_{elem}}$ elements in the domain can be written as follows:
 \begin{align*}
     \begin{split}
-        L_v(W,b) &= \frac{1}{\texttt{N\_elem}}\sum_{k=1}^{\texttt{N\_elem}}\left| \int_{K_k} \left( \nabla u_{\text{NN}}(x;W,b) \cdot \nabla v_k dK  ~ - ~ \int_{K_k} f\,v_k\,dK ~\right) ~\right|^2,\\
-        L_b(W,b) &= \frac{1}{N_D}\sum_{d=1}^{N_{D}}\left|u_{\text{NN}}(x; W, b) - g(x)\right|^2,\\
+        L_v(W,b) &= \frac{1}{\texttt{N\_elem}}\sum_{k=1}^{\texttt{N\_elem}}  \left( \int_{K_k} \nabla u_{\text{NN}}(x;W,b) \cdot \nabla v_k dK  ~ - ~ \int_{K_k} f\,v_k\,dK ~\right)^2,\\
+        L_b(W,b) &= \frac{1}{N_D}\sum_{d=1}^{N_{D}}\left(u_{\text{NN}}(x; W, b) - g(x)\right)^2,\\
         L_{\text{VPINN}}(W,b) &= L_v + \tau L_b.
     \end{split}
 \end{align*}
 
[…]
