### 2024-07-29 2ba848e
JOSS paper release v1.0.2
Changed the version Number
