### 2024-07-29 #6764 @thivinanandh
> @thivinanandh I am fine, please do the following:
>
> - [x]     Double check authors and affiliations (including ORCIDs)
> - [x]     Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> - [x]      Archive the release on Zenodo/figshare/etc and post the DOI here.
> - [x]     Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> - [x]     Make sure that the license listed for the archive is the same as the software license.

Dear @diehlpk,

1, **Authors and Affiliations**

The Authors Names and affiliations within the paper are correct

2 **Release**

A new release of the code is made at with
Github Release Version Number : ```1.0.2```

It can be accessed at [FastVPINNs - v1.0.2 ](https://github.com/cmgcds/fastvpinns/releases/tag/v1.0.2)
Note: The same updated version of the code is also pushed to pypi, `fastvpinns==1.0.2` can be accessed at [PyPI Link](https://pypi.org/project/fastvpinns/1.0.2/)

3, **Zenodo**

The Code with the version `1.0.2` is uploaded to Zenodo with the DOI :
```
https://doi.org/10.5281/zenodo.13127234
```

4, **Title and Author List**

The Title and the Author list from the paper matches with the Zenodo Archive's title and authors.

5, **Licence**

Both Zenodo Archive's licence and the licence of the code submitted to JOSS are set to `MIT` licence.

Please let me know if you need any additional information regarding the same,
