### 2024-07-29 #6764 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
