### 2024-07-30 c3663b6
minor changes in paper.md
