+++ joss/paper.md
@@ -35,7 +35,7 @@ Partial differential equations (PDEs) are essential in modeling physical phenome
     u(x) &= g(x), \quad x \in \partial \Omega.
 \end{align}
 
-Here, $x \in \Omega$ is the spatial co-ordinate, $u(x)$ is the solution of the PDE, $f(x)$ is a known source term and $g(x)$ is the value of the solution on the domain boundary, $\partial \Omega$. 
+Here, $x \in \Omega$ is the spatial coordinate, $u(x)$ is the solution of the PDE, $f(x)$ is a known source term, and $g(x)$ is the value of the solution on the domain boundary, $\partial \Omega$. 
 
 A neural network is a parametric function of $x$, denoted as $u_{\text{NN}}(x; W, b)$. In this context, $W$ and $b$ represent the weights and biases of the network. When the neural network consists of $h$ hidden layers, with the $i^{\text{th}}$ layer containing $n_i$ neurons, the mathematical representation of the function takes the following shape:
 \begin{equation*}
@@ -45,7 +45,7 @@ A neural network is a parametric function of $x$, denoted as $u_{\text{NN}}(x; W
 
 Here, $l : \mathbb{R}^{n_h} \rightarrow \mathbb{R}$ is a linear mapping in the output layer and $\mathrm{T}^{(i)}(\cdot) = \sigma(W^{(i)} \times \cdot + b^{(i)})$ is a non-linear mapping in the $i^{th}$ layer $( i= 1, 2, \cdots , h)$, with the non-linear activation function $\sigma$ and the weights $W^{(i)}$ and biases $b^{(i)}$ of the respective layers.
 
-Physics-informed neural networks (PINNs), introduced by [@raissi2019physics], work by incorporating the governing equations of the physical systems and the boundary conditions as physics-based constrains to train the neural networks. The core idea of PINNs lies in the ability of obtaining the gradients of the solutions with respect to the input using the automatic differention routines available within deep learning frameworks such as tensorflow [@tensorflow2015-whitepaper]. The loss function for the PINNs can be written as follows:
+Physics-informed neural networks (PINNs), introduced by @raissi2019physics, work by incorporating the governing equations of the physical systems and the boundary conditions as physics-based constrains to train the neural networks. The core idea of PINNs lies in the ability of obtaining the gradients of the solutions with respect to the input using the automatic differention routines available within deep learning frameworks such as TensorFlow [@tensorflow2015-whitepaper]. The loss function for the PINNs can be written as follows:
 
 \begin{align*}
     \begin{split}
@@ -55,9 +55,9 @@ Physics-informed neural networks (PINNs), introduced by [@raissi2019physics], wo
     \end{split}
 \end{align*}
 
-Here, the output of the neural network, $u_{\text{NN}}(x; W, b)$, is used to approximate the unknown solution. In addition, $N_T$ is the total number of training points in the interior of the domain $\Omega$, $N_D$ is the total number of training points on the boundary $\partial\Omega$, $\tau$ is a scaling factor applied to control the penalty on the boundary term and $L_{\text{PINN}}(W,b)$ is the loss function of the PINNs. 
+Here, the output of the neural network, $u_{\text{NN}}(x; W, b)$, is used to approximate the unknown solution. In addition, $N_T$ is the total number of training points in the interior of the domain $\Omega$, $N_D$ is the total number of training points on the boundary $\partial\Omega$, $\tau$ is a scaling factor applied to control the penalty on the boundary term, and $L_{\text{PINN}}(W,b)$ is the loss function of the PINNs. 
 
[…]
