### 2024-07-30 #6764 @danielskatz
@thivinanandh - I'm suggesting small changes in https://github.com/cmgcds/fastvpinns/pull/22 - please merge this, or let me know what you disagree with, then we can continue the acceptance process.
