### 2024-07-30 #6764 @thivinanandh
Hi @danielskatz,

Thanks for making the PR to make corrections in the Paper. We did check the commit and we have approved the PR to merge with the `joss_paper` branch. Please let is know if we need to do anything else regarding the same.
