### 2024-07-30 #6764 @danielskatz
Congratulations to @thivinanandh (Thivin Anandh) and co-authors on your publication!!

JOSS depends on volunteers and couldn't be successful without you
